# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "mlvt",
        "version": "1.0.0",
        "contact": {}
    },
    "paths": {
...
        "/videos": {
            "post": {
                "description": "Add a new video for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Add a new video",
                "parameters": [
                    {
                        "description": "Video details",
                        "name": "video",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.AddVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Video added successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/videos/generate-presigned-url": {
            "post": {
                "description": "Generates a pre-signed URL for uploading a video to AWS S3 and registers the video with initial data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Videos"
                ],
                "summary": "Generate a pre-signed URL for video uploads",
                "parameters": [
                    {
                        "description": "Video upload details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.GeneratePresignedURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.GeneratePresignedURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to generate presigned URL or Failed to add video",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/videos/user/{userID}": {
            "get": {
                "description": "Get all videos uploaded by a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get videos by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of videos",
                        "schema": {
                            "$ref": "#/definitions/schema.GetVideosResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/videos/{id}": {
            "get": {
                "description": "Get details of a video by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Get video by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video details",
                        "schema": {
                            "$ref": "#/definitions/schema.Video"
                        }
                    },
                    "400": {
                        "description": "Invalid video ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing video",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Update video details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated video details",
                        "name": "video",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.AddVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid video ID or request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a video by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "videos"
                ],
                "summary": "Delete video",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Video ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid video ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "schema.AddVideoRequest": {
            "type": "object",
            "required": [
                "duration",
                "link",
                "title",
                "user_id"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "title": {
                    "description": "Add Title field",
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "schema.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "schema.GeneratePresignedURLRequest": {
            "type": "object",
            "required": [
                "duration",
                "file_name",
                "file_type",
                "title",
                "user_id"
            ],
            "properties": {
                "duration": {
                    "description": "Expected duration of the video",
                    "type": "integer"
                },
                "file_name": {
                    "description": "Name of the video file to be uploaded",
                    "type": "string"
                },
                "file_type": {
                    "description": "Type of the video file (e.g., video/mp4)",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the video",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID of the user uploading the video",
                    "type": "integer"
                }
            }
        },
        "schema.GeneratePresignedURLResponse": {
            "type": "object",
            "properties": {
                "presignedUrl": {
                    "type": "string"
                }
            }
        },
        "schema.GetVideosResponse": {
            "type": "object",
            "properties": {
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Video"
                    }
                }
            }
        },
...
        "schema.Video": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp of when the video was created",
                    "type": "string"
                },
                "duration": {
                    "description": "Duration of the video in seconds",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "description": "URL to the video on AWS S3",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the video",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Timestamp of the last update to the video",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID of the user who uploaded the video",
                    "type": "integer"
                }
            }
        }
    }
}